+++ TeamPi_6968_DeepSpace/TeamPi_6968_DeepSpace/src/main/include/Adafruit_INA219.h
@@ -212,10 +212,10 @@ class Adafruit_INA219{
   void setCalibration_32V_2A(void);
   void setCalibration_32V_1A(void);
   void setCalibration_16V_400mA(void);
-  float getBusVoltage_V(void);
-  float getShuntVoltage_mV(void);
-  float getCurrent_mA(void);
-  float getPower_mW(void);
+  double getBusVoltage_V(void);
+  double getShuntVoltage_mV(void);
+  double getCurrent_mA(void);
+  double getPower_mW(void);
 
  private:
   frc::I2C *_i2c;
@@ -225,7 +225,7 @@ class Adafruit_INA219{
   // The following multipliers are used to convert raw current and power
   // values to mA and mW, taking into account the current config settings
   uint32_t ina219_currentDivider_mA;
-  float    ina219_powerMultiplier_mW;
+  double   ina219_powerMultiplier_mW;
 
   void init();
   void wireWriteRegister(uint8_t reg, uint16_t value);
